Split, name, & link files from gm_1A4C (#1797)
* Match gm_801A9B30_OnEnter

* gm_1A4C matches

* Split out gmregtyfall, gmmovieend

* Split out gmhowto, gmomake15

* Split out gmopening, gmstaffroll

* Use header-defined extern VS mode data

* Match & link gmvsmelee

* Sort static.h header after main header, but before all others

* Remove now-empty gm_1A4C.static.h

* Fix issues

+++ src/melee/gm/gm_1A3F.c
@@ -175,7 +175,7 @@ void gm_SetSceneMinor(u8 arg0)
     gm_80479D30.nums.prev_minor = arg0;
 }
 
-void gm_SetScenePendingMinor(int pending_minor)
+void gm_SetScenePendingMinor(u8 pending_minor)
 {
     gm_80479D30.nums.pending_minor = pending_minor + 1;
 }
